feat: use typescript not jsx
