+++ src/components/icons/check.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
-import { getSizeClass, getColorClass } from './iconUtils';
+import React, { FC } from 'react';
+import { getSizeClass, getColorClass, IconProps } from './iconUtils';
 
-const CheckIcon = ({ size, color }) => {
+const CheckIcon: FC<IconProps> = ({ size, color }) => {
 	const svgClasses = getSizeClass(size);
 	const pathClasses = getColorClass(color);
